+++ src/commonMain/kotlin/csense/kotlin/annotations/threading/InAnyContext.kt
@@ -2,6 +2,10 @@
 
 package csense.kotlin.annotations.threading
 
+/**
+ * Denotes that the given context (the whole class/file etc.)
+ * can execute in whatever threading state (say UI, background, IO etc)
+ */
 @Retention(AnnotationRetention.BINARY)
 @Target(
     AnnotationTarget.CLASS,