+++ src/commonMain/kotlin/csense/kotlin/annotations/threading/InBackground.kt
@@ -6,7 +6,6 @@ package csense.kotlin.annotations.threading
  * Denotes that something MUST happen in the background(threading)
  * so eg IO (should not be done in the UI)
  * or long computing tasks / slow computations.
- *
  */
 @Target(
     AnnotationTarget.FUNCTION,
@@ -15,6 +14,6 @@ package csense.kotlin.annotations.threading
     AnnotationTarget.PROPERTY_GETTER,
     AnnotationTarget.VALUE_PARAMETER
 )
-@Retention(AnnotationRetention.BINARY) //binary so its persisted across libraries
+@Retention(AnnotationRetention.BINARY) 
 @MustBeDocumented
 public annotation class InBackground