+++ src/commonMain/kotlin/csense/kotlin/annotations/threading/InBackgroundContext.kt
@@ -2,6 +2,12 @@
 
 package csense.kotlin.annotations.threading
 
+/**
+ * Denotes that something MUST happen in the background(threading)
+ * so eg IO (should not be done in the UI)
+ * or long computing tasks / slow computations.
+ * This applies to the whole scope (eg a whole class)
+ */
 @Retention(AnnotationRetention.BINARY)
 @Target(
     AnnotationTarget.CLASS,